+++ tortoise/tests/test_generate_schema.py
@@ -5,7 +5,7 @@ from asynctest.mock import CoroutineMock, patch
 from tortoise import Tortoise
 from tortoise.contrib import test
 from tortoise.exceptions import ConfigurationError
-from tortoise.utils import get_schema_sql
+from tortoise.utils import generate_post_table_sql, get_schema_sql
 
 
 class TestGenerateSchema(test.SimpleTestCase):
@@ -44,9 +44,15 @@ class TestGenerateSchema(test.SimpleTestCase):
                 }
             )
             self.sqls = get_schema_sql(Tortoise._connections["default"], safe).split("; ")
+            self.post_sqls = generate_post_table_sql(Tortoise._connections["default"], safe).split(
+                "; "
+            )
 
     def get_sql(self, text: str) -> str:
         return re.sub(r"[ \t\n\r]+", " ", [sql for sql in self.sqls if text in sql][0])
+
+    def get_post_sql(self, text: str) -> str:
+        return re.sub(r"[ \t\n\r]+", " ", [sql for sql in self.post_sqls if text in sql][0])
 
     async def test_noid(self):
         await self.init_for("tortoise.tests.testmodels")
@@ -132,6 +138,14 @@ class TestGenerateSchema(test.SimpleTestCase):
         ):
             await self.init_for("tortoise.tests.models_m2m_1")
 
+    async def test_table_and_row_comment_generation(self):
+        await self.init_for("tortoise.tests.testmodels", safe=True)
+        self.continue_if_safe_indexes(supported=False)
+        sql = self.get_sql("comments")
+        self.assertRegex(sql, r".*\/\* Upvotes done on the comment.*\*\/")
+        self.assertRegex(sql, r".*\\n.*")
+        self.assertRegex(sql, r".*it\'s.*")
+
 
 class TestGenerateSchemaMySQL(TestGenerateSchema):
     async def init_for(self, module: str, safe=False) -> None:
@@ -179,6 +193,15 @@ class TestGenerateSchemaMySQL(TestGenerateSchema):
         )
         self.assertIn("`team_id` INT NOT NULL REFERENCES `team` (`id`) ON DELETE CASCADE", sql)
 
+    async def test_table_and_row_comment_generation(self):
+        await self.init_for("tortoise.tests.testmodels", safe=True)
+        self.continue_if_safe_indexes(supported=False)
+        sql = self.get_sql("comments")
+        self.assertIn("COMMENT=", sql)
+        self.assertIn("COMMENT ", sql)
+        self.assertRegex(sql, r".*\\n.*")
+        self.assertRegex(sql, r".*it\\'s.*")
+
 
 class TestGenerateSchemaPostgresSQL(TestGenerateSchema):
     async def init_for(self, module: str, safe=False) -> None:
@@ -209,3 +232,16 @@ class TestGenerateSchemaPostgresSQL(TestGenerateSchema):
         sql = self.get_sql('"noid"')
         self.assertIn('"name" VARCHAR(255)', sql)
         self.assertIn('"id" SERIAL NOT NULL PRIMARY KEY', sql)
+
+    async def test_table_and_row_comment_generation(self):
+        await self.init_for("tortoise.tests.testmodels", safe=True)
+        self.continue_if_safe_indexes(supported=False)
+        sql = self.get_post_sql("comments")
+        self.assertIn("COMMENT ON TABLE", sql)
+        self.assertIn("COMMENT ON COLUMN", sql)
+        self.assertRegex(sql, r"COMMENT ON TABLE comments IS.*")
+        self.assertRegex(
+            sql, r"COMMENT ON COLUMN comments\..* IS.*;.*COMMENT ON COLUMN comments\..* IS.*"
+        )
+        self.assertRegex(sql, r".*\\n.*")
+        self.assertRegex(sql, r".*it\'s.*")